+++ jsapp/js/components/submissions/table.es6
@@ -62,6 +62,7 @@ import tableStore from 'js/components/submissions/tableStore';
 import './table.scss';
 import MediaCell from './mediaCell';
 import AudioCell from './audioCell';
+import {isSubmissionWritable} from 'js/components/permissions/utils';
 
 const DEFAULT_PAGE_SIZE = 30;
 
@@ -415,9 +416,9 @@ export class DataTable extends React.Component {
                 checked={this.state.selectedRows[row.original._id] ? true : false}
                 onChange={this.bulkUpdateChange.bind(this, row.original._id)}
                 disabled={!(
-                  (this.isSubmissionWritable('change_submissions', this.props.asset, row.original)) ||
-                  (this.isSubmissionWritable('delete_submissions', this.props.asset, row.original)) ||
-                  (this.isSubmissionWritable('validate_submissions', this.props.asset, row.original))
+                  (isSubmissionWritable('change_submissions', this.props.asset, row.original)) ||
+                  (isSubmissionWritable('delete_submissions', this.props.asset, row.original)) ||
+                  (isSubmissionWritable('validate_submissions', this.props.asset, row.original))
                 )}
               />
             }
@@ -431,7 +432,7 @@ export class DataTable extends React.Component {
               <i className='k-icon k-icon-view'/>
             </button>
 
-            {userCanSeeEditIcon && (this.isSubmissionWritable('change_submissions', this.props.asset, row.original)) &&
+            {userCanSeeEditIcon && (isSubmissionWritable('change_submissions', this.props.asset, row.original)) &&
               <button
                 onClick={this.launchEditSubmission.bind(this)}
                 data-sid={row.original._id}
@@ -499,7 +500,7 @@ export class DataTable extends React.Component {
         <ValidationStatusDropdown
           onChange={this.onValidationStatusChange.bind(this, row.original._id, row.index)}
           currentValue={this.getValidationStatusOption(row.original)}
-          isDisabled={!(this.isSubmissionWritable(PERMISSIONS_CODENAMES.validate_submissions, this.props.asset, row.original))}
+          isDisabled={!(isSubmissionWritable(PERMISSIONS_CODENAMES.validate_submissions, this.props.asset, row.original))}
         />
       ),
     };
